Change buck2 test command to directly use tokio spawn
Summary:
As of D45634598, we are making temporary_spawn only provide a reference to DiceComputation.
This means that the use of temporary spawn for tests will no longer provide the DiceTransaction that it intends to receive to store in the test orchestrators.

We can receive the DiceTransaction directly from the command, and then spawn using regular tokio and dupe the transaction.

Reviewed By: krallin

Differential Revision: D45834160

fbshipit-source-id: 39e70f5543e0fcc8c7ad054d5aa0f0f74c954870