Make sure termination observer is notified as complete when future completes normally
Summary:
There's an issue where TerminationObserver would only resolve to Poll::Ready if we called `cancel`, even though we can obtain one without ever calling `cancel`.
This is misleading and bug prone because the result of this even has different termination status that includes `Ready`, which means that this should resolve when the future is actually ready.

So, make the termination observer always resolved when the actual future is resolved, or cancelled, or just dropped.

Reviewed By: krallin

Differential Revision: D45801521

fbshipit-source-id: aea8976a48c16599db664130ca000b8cdea5da0c