buck2: span recording: record on first poll, not creation
Summary:
Like it says in the title.

This ensures that doing `async_record_spans(span_async(...))` actually works
even though the span was created before we wrapped it.

Note that it seems debatable whether we should be emitting the span before the
first `poll()` (which right now we do). Right now, we are. Perhaps we
shouldn't. That's a separate problem however.

Differential Revision: D45866774

fbshipit-source-id: 199f562c97d838698a8ea205485cdf7d39a165e5