buck2: critical path: use DICE activations
Summary:
This switches up the notifications to our critical path listener to go through
DICE instead of via a separate side channel that gets written into as part of
evaluating a key. The upsides of doing it this way are:

- We don't get duplicate notifications due to eager deps. Arguably, that problem
  will be fixed in time.
- We do get notifications when early cutoff means a node doesn't get evaluated
  at all. This one needs a solution (see the attached task).

This diff explains what an "activation" is: D45602534

This has some downsides, namely that we need to have a little more logic
dedicated to understanding the DICE graph here, and we must keep track of a
few more keys.

The tradeoff here is that we have something that is a bit harder to get right for
the simple cases (because you need more keys), but if it works in the simple
cases then it also works in the hard cases (eager deps, early cutoff).

(I think that's worth it)

There is also a bit of refactoring here, which consists of moving the calls to
set the activation data up the stack (closer to the key). The previous enables
doing this easily by letting us capture all the (top level) spans that get
involved in the evaluation of a key.

Finally, this moves the keys to be public, since we need to downcast into them.
I know we've historically tried to keep DICE an implementation detail, but I'd argue
that's generally a bit of mistake (and in particular in cases like this): we need
to know what DICE is doing for things like this critical path to work.

Reviewed By: bobyangyf

Differential Revision: D45702825

fbshipit-source-id: e9a0944a7c3b8a2ca14fe3d82e5435be6d4b1e95