buck2: add a mechanism to force a daemon restart via a version file in the repo
Summary:
For the Blake3 migration, we are going to want to ensure that we have one
atomic commit where all clients before this use SHA1, and all clients after
use Blake3.

However, if we landed that config change today, it wouldn't work, because we
would not restart the daemon when the config changes, so you'd get SHA1 or
Blake3 depending on which side of the commit your daemon started on.

This diff allows us to fix that by introducing support for a "daemon busting"
mechanism, which is a file you put in the repo that allows us to force the
client to deem the daemon incompatible if it doesn't match. This is read by the
client on every invocation, and passed to the daemon at startup.

A more ideal approach here would be to actually have a mechanism to have
dedicated configs for the daemon that result in a restart if they mismatch.
However, that's a bigger change, and one that requires quite a bit more design
upfront, so I'd like us to make progress on the Blake3 migration first.

Besides, having this busting mechanism seems *generally useful* in case we need
to convert a "restart your daemon" recovery instruction into a "rebase past
this commit" (which we have needed in the past), so that seems worth having.

Reviewed By: get9

Differential Revision: D45903432

fbshipit-source-id: 39b74e8637ee53818d6c2802d480bd1b95ea9692