buck2: cleanup_path: use fs_util
Summary:
Like it says in the title. This doesn't need to do its own call in order
to see the one error kind it cares about.

This, like the previous diff, lets us keep calls to `as_maybe_relativized` in
`fs_util`.

Note that I'm replacing a call to `fs.remove_path_recursive(path)?;` with one
to just `fs_util::remove_file`, but if you look at what `remove_path_recursive`
does, it's exactly what it does when the file is not a directory.

Reviewed By: themarwhal

Differential Revision: D45903919

fbshipit-source-id: 93f17cd3e7c00e793402a4fe05685caa0a0e3753