buck2: re: don't convert paths to String
Summary:
Like it says in the title. This is a bit of cleanup that enables diffs later in
this stack.

Note that there is a bit of a behavior change: we now pass an absolute path
as the "meterialization_mount_path" (sic). The RE client ends up stat-ing
this path, which his fine (but that actually revealed a potential bug later
in this stack).

Reviewed By: stepancheg

Differential Revision: D45864516

fbshipit-source-id: a3e5fffb44704b410d18689cd54684d2465c2e1e