buck2: add support for changing directories to buck-out and relativizing paths
Summary:
Like it says in the title. This adds support for cd-ing to buck-out

We do this so that we don't have to do any path traversals through (a
potentially slow) NFS mount that the repository might be on. See the
attached task for details.

This is, in effect, a poor man's `openat`: we just use the current working
directory as our one "at" point.

Differential Revision: D45820046

fbshipit-source-id: 6d7d3a83418ca4168db7e3fd226ed2d6d4928a31