Add VersionEpoch and thread it through dice for computing keys
Summary:
Add a VersionEpoch that is incremented everytime a new instance of a SharedComputationCtx is created.

This is used so that in D45925966 we can track which instance of the shared context each task is running for, so that when we cancel all tasks when a SharedComputationCtx is no longer referenced (D45634597), we consider that VersionEpoch no longer relevant and ignore any cache updates from it.

Reviewed By: stepancheg

Differential Revision: D45925965

fbshipit-source-id: c22d43ce144f7e3971d8594fe9645da1806bdaac