Drain active tasks from evicted shared state
Summary:
on dropping the shared state, which is when the top level request goes out of scope, we cancel all tasks in DICE since there is no more active computations.

We return the TerminationObserver for them so that we can observe when dice becomes idle

There is a better idea that instead of doing wait_for_idle, we can, since getting a transaction is now async, we can have dice automatically block handing out new transaction until state is cleaned up. But that's a bigger change to command + dice handling that should be done separately, independent of efforts to bring up modern dice

Reviewed By: stepancheg

Differential Revision: D45634597

fbshipit-source-id: 8a4eb5d4d3c3ad4d08d1d326cd5965dbff9a7bd6