Allow unused on some useful function on key types
Summary:
`DiceKeyErased` can be created from `key` or `proj`.
We happen to not use `proj` because in real code, we create it as `DiceKeyErasedRef` instead. In tests, we do use this constructor.

anyhow, having this function is consistent, and doesn't bring harm, so will just allow it to be unused

Reviewed By: stepancheg

Differential Revision: D45717494

fbshipit-source-id: d12c936ab5dd17791fe103d9ab75abc9c6b88411