buck2_artifact crate
Summary:
`buck2_artifact` contains `Artifact` type and dependencies.

The crate is not very large, but it does not pull heavy dependencies except `dice`:

{F994194240}

(`dice` is not needed, and we should fix that too).

This diff is good as is, but it is part of the plan of pulling providers from `buck2_build_api`. Providers depend on artifacts.

Reviewed By: krallin

Differential Revision: D45931382

fbshipit-source-id: ee95e8e258388b03dd7a850405447d4eccc80677