Simplify kill logic
Summary:
Do not use fancy (and good!) windows API where we create process handle and wait on it for termination.

Instead do replicate dumb Unix logic: terminate the process, and then check if it still exists ten times per second.

This should make code easier to maintain: code path of retries and timeout is the same on Unix and Windows.

Add unit tests for two functions: `kill_impl` and `process_exists`.

This is done to use the same kill code for `kill` and `killall` (D45976291).

Reviewed By: themarwhal

Differential Revision: D45975504

fbshipit-source-id: 69ca83a12d0e69d448e1d0c40efac6991305745b