Add introspection of DiceTask state for dice dumps.
Summary:
We can now use the internal state of dice tasks to make a dump.
This was some bits left over to fully integrate dice dumps since we didn't have
the new futures fully implemented.

Now the current situation isn't ideal in the mapping of states.
We should improve the states that we output here as dice tasks have much more
fine grained information that what is currently being produced for dice dumps.

That will come later. The priority right now is to hook everything together.

Reviewed By: stepancheg

Differential Revision: D45964585

fbshipit-source-id: f382c27659f20edef4800fc18c8b47bde09bad86