Fix dice future where a finished cancelling task will not return cancelled
Summary:
Fix bug in modern dice where it would get errors saying `key was cancelled`

This is a bug where a task that was cancelled would not return `MaybeCancelled::Cancelled` if it finished cancelling. Instead it returned a valid value.
This made the dice shared cache read the value and think that was a legit value instead of a cancelled value.

Longer term:
I think having DiceError::cancelled is a bad variant. Cancelled should be treated specially in dice rather than just an error that is user visible (its actually not expected to be so anyways).
I think we need to refactor the dice result type into user visible vs internal so that we distinguish errors vs some cancelled future, but that can come later

Reviewed By: stepancheg

Differential Revision: D45930894

fbshipit-source-id: 08ed6af4546b959b24be8ce13479907ef39c16a7