buck2: move build signals behind a trait
Summary:
Like it says in the title. My end goal here is to allow moving this code out
of build API so it can build faster, because waiting on build API when making
changes here is rather annoying.

To do this, change the places that use BuildSignalsListener to do so via a
trait. This means adding the trait, but also doing the conversion to the DICE
ActivationTracker earlier since you can't cast a `Arc<dyn X>` to a `Arc<dyn
Y>`.

Reviewed By: bobyangyf

Differential Revision: D45947557

fbshipit-source-id: 8656eb82054ee12161fd04c8b4c6db73d00032e2