buck2: move build signals to their own crate
Summary:
This makes it possible to work on this code without rebuilding all of
buck2_build_api, which is nice.

In time, if the various DICE keys we use in buck2_build_api also move out of
that crate, we'll be able to have this not depend one buck2_build_api at all.

Reviewed By: bobyangyf

Differential Revision: D45947555

fbshipit-source-id: cac454d70ca6f892a1bb3d0da7ada2c2330b6d9f