adding the possibility for test executor to show info messages to users 1/6
Summary:
In this stack we had the possibility for the test executor to show info messages to users when they run tests with buck2. The messages are shown at the end of the run before the buck test command finishes and are meant to help users or teach them about features in some very specific cases (WE DON'T WANT THESE MESSAGES TO BE SPAMMY).

Specifically, we want the following 2 messages coming from tpx to be shown:
- how to debug a failing test with fdb. This messages is shown ONLY when there is a failing test and provides the command to start debugging
- how to visualize local coverage in VS Code. This messages is shown ONLY if the user enabled coverage collection (passed --collect-coverage to tpx) and testify reporting is working (the user is in fbsource and uses VS Code).

To implement this feature we added a new API call in the test orchestrator that allows the executor to attach a info message.

Reviewed By: krallin

Differential Revision: D45775342

fbshipit-source-id: 1f5b4fb5d417a5b94422b5c985c3f01f6ae89cbe