buck2: default tests to run from project root & with relative paths in OSS
Summary:
Those were concessions we made internally for compatibility with Buck1, we
should probably not default to the tech debt for OSS users.

This is a breaking change, but it probably makes sense to do it earlier than later
while we don't have too many users.

See: https://github.com/facebook/buck2/issues/264

Reviewed By: themarwhal

Differential Revision: D46020229

fbshipit-source-id: 53e613eef8699b0aaa0a7aea126453812fdb304e