Make BxlKeyData display less verbose
Summary:
ActionKey is made up of `owner()` display. `BxlKeyData` prints out every single CLI arg, so we get gRPC errors with metadata size.

krallin mentioned that ActionKey on RE Is used for memory predictions. The preference is a lower cardinality, so the BXL script owner can and should make the category or identity more unique if they want to get back higher cardinality. Let's just populate the action key with the `BxlFunctionLabel` instead.

Reviewed By: krallin

Differential Revision: D46024481

fbshipit-source-id: abd2881e74b24d331870788be451e656b9a12cc5