buck2: move more configs into DaemonStartupConfig
Summary:
Like it says in the title. This allows us to automatically restart the daemon
when those configs change.

For the digest algorithm config, that's necessary to ensure we correctly handle
rebasing past the diff that changes the config.

For cwd_buck_out, that'll allow us to roll this out via a GK.

Reviewed By: stepancheg

Differential Revision: D46021398

fbshipit-source-id: 7aa50a92cb9efc32e01a07f210aa39d9cd32fcea