Extract compression from MultiDexExecutableMain
Summary:
Previously, we had a single command `MultiDexExecutableMain` that would (a) create a primary dex (optionally) and N secondary dexes (also optionally), and (b) process the secondary dexes to compress them and add metadata etc.

That command was too big and unwieldy to understand. Instead, split it into separate (a) and (b).

Sorry, this is quite a big diff but all the changes need to land simultaneously to retain existing behavior.

Reviewed By: ianlevesque

Differential Revision: D46067744

fbshipit-source-id: dc5273ce40a3fe22e1930597876e17f295264ca4