Add worker id to worker request
Summary:
Adding an id to WorkerSpec to uniquely identify a type of worker, so that workers are shared for a given `WorkerInfo` at a given state. This is used as the key in a hash map of workers, as well as the path to the UDS socket in the next diff.

For this purpose it doesn't matter if the id outlives the underlying starlark value, hence using `unsafe` to get a `ValueIdentity` with a static lifetime.

Reviewed By: cjhopman

Differential Revision: D45934283

fbshipit-source-id: bdc279f52784d426e1a293ff03e1178abadd7787