Add result type for dice
Summary:
Make cancelations a separate type
They are really quite special and should be explicit

This is an result type for now, but really I would like it to be it's own Enum of OK and Cancelled.
However, it gets quite verbose as it is nice to be able to propagate them via '?'. That requires unstable feature 'try_v2' so that we can implement 'Try' for custom types. The difference doesn't really warrant using an unstable feature.
I could change this to an enum and propagate explicitly instead of via '?' though.

Ultimate goal is to resolve `Key was cancelled` errors and the panic attached

Differential Revision: D45970867

fbshipit-source-id: c418cf45b37947fb032cb064960c7e6f090dce7d