Make eval_entry_versioned return CancellableResult
Summary:
Make this return the result, then have the caller set the dice task handle with the result

This makes it clear the guarantee around when the result is set vs canceled.
Now we only have 1 place where we set the result instead of 5 different branches

This adds a todo which is the cultprit of the panic task attached.

Reviewed By: stepancheg

Differential Revision: D46019141

fbshipit-source-id: e1d596b5b1475159f9f9ae987f9a170bc32ce023