Revert D46012365: Multisect successfully blamed D46012365 for test or build failures
Summary:
This diff is reverting D46012365
D46012365: [fbcode_macros][native python] Support building native python w/ PIE by andrewjcg has been identified to be causing the following test or build failures:

Tests affected:
- [aiplatform/training/adapters/tests:state_dict_adapter_tests - aiplatform.training.adapters.tests.state_dict_adapter_tests.StateDictAdapterTests: test_apply_rules_to_stateful_module](https://www.internalfb.com/intern/test/844425007745950/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2133255
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: yixin94

Differential Revision: D46220131

fbshipit-source-id: 0e63544c0af5470d5ce9629ba1a0497deab68876