Add warning if buck2 run is used with multiple targets
Summary:
A user tried to execute `buck2 run` with a recursive target expression. Normally, recursive target expressions would cause an error, since BUCK files usually will contain a non-runnable target :

```
Command failed: Target `cell//path/to/some/random/target` is not a binary rule (only binary rules can be `run`)
```
But sometimes, if the BUCK file only contains runnable targets, then the `buck2 run` command will only run the first one that it found, and not tell the user that this happened.

Let's add a soft error here - I'm not sure if we want to support running multiple targets, but we should at least notify the user that this is the current supported behavior.

Reviewed By: stepancheg

Differential Revision: D46279310

fbshipit-source-id: eb9ebac74a4e4584c615e66570dea642c784b709