Fix time flame profiler
Summary:
I incorrectly removed instrumentation for profiler in this stack D46072685.

After cjhopman switched statement profiling to custom profiler, I incorrectly assumed instrumentation is not needed anywhere. But it was needed for heap profiling.

This diff removes the need for instrumentation for heap profiling.

However, now only `def` calls are intercepted, not native calls. This is probably fine.

Reviewed By: bobyangyf

Differential Revision: D46316832

fbshipit-source-id: 19ffead71af5c6ca81b4379888d105d30358c51a