Extend custom error formatting protocol to property maps
Summary:
We extend the protocol for custom formatting of errors, so that the error properties where the formatter is specified can be not only a proplist, but also a map.

Maps are more convenient for writing properties; can be given precise types in specs; and the formatter function already expects the properties to be a map!

Reviewed By: Balajiganapathi

Differential Revision: D46336315

fbshipit-source-id: 08cce96e95d7c307404a1a20d94bce68d751433f