Make includes explicitly relative
Summary:
## This Stack:
With Xcode 14.3+, Apple changed the behavior of clang and it's indexing such that we're required to provide absolute paths to clang. Even when the command succeeds with relative paths, Xcode is not able to provide completions. After extensive testing in S343754, the only means to get Xcode to successfully index was to provide absolute paths. This applies to all args and header maps.

Of course, ***we never want to build with absolute paths***, as each user would capture their repo path in the command and would invalidate all caches. As such, all our compile/link commands execute from the repo root and utilize relative paths.

This allows us to create separate argsfiles and headermaps that utilize absolute paths. They will be output as a separate subtarget and ***never used in actual build***. However, to make sure we update all args correctly, we need to do this alongside our existing generation of header maps and args propagation.

## This diff
To make absolute paths in the argsfiles, we need to replace `./` in args with the user's project root (referred to as `absolute_path_prefix` in the stack. To make this easier, let's add make the fact that these paths are relative explicit.

***Note***: From a compiler perspective, this should be a no-op for all compilers. However, it will invalidate all compile actions as argsfiles are different. The output though should be the same.

Reviewed By: narissiam

Differential Revision: D46334204

fbshipit-source-id: c580e944e9ad7c995373d306686c1ecbe52760b0