Extract out CPreprocessorArgs
Summary:
In the next diff, we begin to compute absolute path args and propagate them up.

To make it clearer to consumers on the expected differences, extract out `CPreprocessorArgs` which are then explicitly referenced as `relative_args`.

We can then add add `absolute_args` safely and in an easier to manage API.

Reviewed By: milend

Differential Revision: D46355437

fbshipit-source-id: 8e1e2c7df5b781141aebad9ea2d93a87b5db6d82