Propagate absolute args in CPreprocessor
Summary:
## For context on this stack, see the detail in D46334204

## This diff:
In the previous diff, we add the ability to generate absolute header maps. In this diff, if we're provided an `absolute_path_prefix`, we generate absolute header maps and args, keeping them separate from the relative ones that are used for build.

We will only perform these actions when `absolute_path_prefix` is set, and we intend to only do so inside an Xcode IDE.

Reviewed By: milend

Differential Revision: D46334206

fbshipit-source-id: 0b805be080f21e1efc1893a51fce85e3d5c48333