Generate absolute argsfiles
Summary:
## For context on this stack, see the detail in D46334204

## This diff:
In D46334206 we ensure that all absolute args are propagated. Now, we're able to ensure that when generating argsfiles, we can also generate absolute argsfiles for use in Xcode.

To do so, we only generate these argsfiles if `absolute_path_prefix` is set. We also perform a regex replacement of `./` with the provided `absolute_path_prefix` (we updated these relative paths in D46334204).

The absolute argsfiles is named `{extension}-abs.argsfile`.

Reviewed By: milend

Differential Revision: D46325904

fbshipit-source-id: f735b5727d6d524705321cc818b6db893a6c2302