Tweak next_gc_level
Summary:
GC is configured incorrectly.

Consider this scenario. Alive set is 1G.

Then we call a function which allocates 1M.

Previous GC threshold was set to 1G + 100K.

So run GC, which keeps alive 1G again and set threshold again to 1G + 100K.

So we constantly run GC which collects almost nothing.

This does not affect regular buck evaluation, because typically when we run GC, nothing is kept alive.

But this affects time flame profiler, which has relatively large alive set.

This makes `time-flame` much faster and fixes the issue with time-flame showing most of time spent in "None" (which is module scope).

Reviewed By: ndmitchell, JakobDegen

Differential Revision: D46376056

fbshipit-source-id: d9a95f95651ac9618ff1055e7f30c34f6e62de5c