Hide Value::new_int from public API
Summary:
On 32 bit `Value` likely needs to be 32-bit. It is smaller.

But also it's hard to make it 64.

For example, in many places we do `Value::get_ref` (which returns `AValueDyn`) which converts `i32` to `&PointerI32`. Which is obviously not possible if `Value` is 64 bit and full `i32` is stored inline.

One possible solution to deal with it, is to store `i29` inline on 32 bit.

(Another solution is to do special case for `i32` in `AValueDyn`).

Not decided yet which approach is easier, but let's hide `Value::new_int` for now to:
* make it easier to experiment
* remove public guarantee that full `i32` is stored inline to be able to change later if needed

Reviewed By: ndmitchell

Differential Revision: D46381356

fbshipit-source-id: 3d2ac64269ce7fe6e5c3772418b65867d83499d0