buck2: maybe_relativize: return `.` when path == cwd, not ``
Summary:
Like it says in the title. This ends up breaking things on Windows if your
buck-out is empty.

In particular, the clean_output_paths walk will walk up the directory tree and
repeatedly get `ENOENT` (this is expected, but on UNIX that would lead us to
exit early so we don't hit the bug there), until we get to `buck-out\v2` which
does exist, but is also our cwd, at which point we'll query the metadata for
``, which is an error, which will lead us to recurse further to `buck-out`
(which is above our cwd), which we'll find exists, and we'll try to remove it
(and succeed).

This paste shows the issue (using the logs from the previous diff: P760102983).

Reviewed By: wendy728

Differential Revision: D46437567

fbshipit-source-id: 98a5b95fa24d3ba6dcb2033733ab9434b402f91c