buck2: make interaction of cwd & relative exe consistent between UNIX & Windows
Summary:
If you do `Command::new("a").current_dir("b")` with your cwd being `c`, on UNIX
you run `c/b/a`, on Windows you run `c/a`.

This is a bit inconsistent, and it means that running Buck2 with a cwd that
is not the project root does not work on Windows.

This diff fixes that by making the behavior consistent on both platforms. When
we run a command, we now check if the executable does exist relative to the
cwd, and if it does, we use it. We know we always provide a cwd and it's always
absolute, so this is fairly robust.

Reviewed By: KapJI

Differential Revision: D46440477

fbshipit-source-id: f2c3bf6877de02fe5e783d3a434e417aa6f46f1f