Handle starting the same key again
Summary:
This was originally written assuming that we can never restart the same key, but that's not true due to cancellations and eager deps.
We can cancel the key at any point, and then start it again, so it's important that the cycle detector can handle these restarts.

Additionally, modern dice and cancellations is written such that there is more cancellation points, and more robust reporting of key finishes even when cancelled, so this is triggered significantly more than in legacy dice (legacy dice would fail to report finished sometimes when cancelled). But anyhow, the exact situation actually doesn't matter whether we report finished or simply start a key twice, the important thing is that we need to restart the node for it due to eager deps.

krallin actually discovered this originally when attempting https://www.internalfb.com/diff/D45521307?dst_version_fbid=1591160794702789
where he ensured that we always report finished.

Reviewed By: stepancheg

Differential Revision: D46210529

fbshipit-source-id: d52f8dab2968c98e57085f7d05b33f20f55f19e1