Refactor connecting to existing_only daemon
Summary:
The flow for existing_only is very different than for the
configurations that will start the daemon. This splits that out so
that we have separate top-level functions for those two flows. The
existing_only flow also can be done with just a daemon dir, it doesn't need the
isolation paths explicitly (and this will enable us to use it for all
daemon dirs for `buck2 status --all`).

Reviewed By: wendy728

Differential Revision: D46041824

fbshipit-source-id: 081f878d625ff9c92dec1a3b092c5fb6c8e0e07d