Waiting on buck2 daemon show trace id
Summary:
```
Waiting on buck2 daemon...
```

message is shown when there are no spans which means:
* rare condition when one span finished and another not yet started
* we forgot to track some important span
* buck2 does nothing and waits for something

First is improbable, and the remaining two are quite important, so print trace id so event log can be quickly fetched by `buck2 log show --allow-remote --trace-id XXX`.

Technically trace id can be found in logs above, but it is not always trivial because logs are often mixed randomly.

Technically trace id can be found by sandcastle id, but that is not always easy if sandcastle ran multiple buck commands.

Reviewed By: JakobDegen

Differential Revision: D46507938

fbshipit-source-id: efe47031b54003af619af0dc56d907bdf1c97f4f