Remove some temporary spawns
Summary:
Prepare to kill `Dice::temporary_spawn`

temporary_spawn makes dice complicated because it makes us copy/clone the DiceComputations, which is not right.
The internal data has to be wrapped in arc, and causes unsafe code like
https://www.internalfb.com/code/fbsource/[74ae4f9ce91fbe0be82144309b0151f9b711b2c2]/fbcode/buck2/dice/dice/src/impls/ctx.rs?lines=347

Additionally, it's preventing some type based assertions on drop order due to the lifetime requirement of static when we support this api through spawning.

Reviewed By: stepancheg

Differential Revision: D46410142

fbshipit-source-id: 0257d94090967fc51ce8c936114ba836d1dc9886