configure tests independently
Summary:
when buck2 test command is invoked.

Imagine we have:
```
my_rule(
 name = "A",
 tests = ["test_of_A"],
 default_target_platform = ":x",
)
my_test(
 name = "test_of_A",
 default_target_platform = ":y",
)
```
We can execute tests as
```
buck2 test :A
```
or as:
```
buck2 test :test_of_A
```
Current behavior is that in the first case `test_of_A` will be built with `:x` platform while in second case it will be built with `:y` platform.

In this diff change how target platform is selected for tests when `buck2 test`
command is run. Select target platform independently for every test from
`tests` attribute independently. New behavior is hidden behind `buck2.independent_tests_platform_resolution` buckconfig for a safer rollout.

Note that cquery behavior is not changed in this diff and will be changed
independently in a follow-up.

Reviewed By: krallin

Differential Revision: D46519836

fbshipit-source-id: e62b855f124a07b2e9aeafb52f31d8a1061f7cd6