Add git_fetch rule to download commit from git repository
Summary:
```
git_fetch(
    name = "serde-rs/serde.git",
    repo = "https://github.com/serde-rs/serde",
    rev = "1aebdc2760a7772c0fdc44329acb3409fc42395e",
    visibility = [],
)
```

Conceptually similar to `http_archive`, but for git.

The `rev` is **required** and is checked after fetch, ensuring that the fetched content is a deterministic function of the attributes, if you trust SHA-1 and git.

I plan to use this to make Reindeer support Cargo's `[dependencies] ... = { git = "..." }` git dependencies.

Reviewed By: diliop

Differential Revision: D46583907

fbshipit-source-id: b729ca15164d4e9bb8f039fa7e63086fb90ea059