buck2: tcp executor: don't hang if the child dies
Summary:
If the child dies before connecting, we just sit there waiting for it to
connect. I don't actually know if that happens (I suspect it might on Windows
with the internal test executor), but in any case we should just guard against
it.

This diff does that by monitoring for the executor exiting while we wait.

Reviewed By: stepancheg

Differential Revision: D46558994

fbshipit-source-id: 9a3c1b3f45f9ba4ddaa881352276cb62711a98ed