LockFreeRawTable for SharedCache
Summary:
When a key is computed, move it into ` LockFreeRawTable` so:
* lookup is cheaper
* release memory held by `CancellationHandle`

This is a bit suboptimal, because:
* `DiceTaskInternal` stores two fields which are no longer needed after value is computed (`state` and `dependants`), this is relatively easy to fix
* there should be one map instead of two, still executing tasks should probably be accessed without locking too. But that is somewhat harder to implement.
* code is a bit complicated

Commented out one test which updates already computed keys. I could not understand what the test is doing, so I could not fix it. We should not write tests downgrading values anyway. And we should write easier to read tests.

Reviewed By: bobyangyf

Differential Revision: D45646301

fbshipit-source-id: f8401b4b45ddfe63ebc5564db1f62c9fc28fc374