resolve_idents_in_type_expr
Summary:
Extract a function.

Following diff D46587577 uses is to validate type expressions more strictly.

Currently `TypeExpr` is a wrapper for `Expr` which is validated at parse time. Ideally, `TypeExpr` should be independent type, I tried, that's a lot of work:
* half of uniplate needs to be rewritten
* a lot of analysis in compiler
* a lot of lints

So let's keep AST as it is for now.

Reviewed By: ndmitchell

Differential Revision: D46587242

fbshipit-source-id: a9b532dc140470ea985a122fff4801ea6666b7e5