buck2: java: expose classpath subtarget
Summary:
This way you can request materialization of the binaries referenced by `audit
classpath`. Maybe it'd be extra convenient if we wrote the paths to a file (and
then the file was an index to find the jars?), but for now let's at least
allow you to ask for their materialization.

See: https://fb.workplace.com/groups/buck2users/permalink/3437668693156154/

Reviewed By: IanChilds, asp2insp

Differential Revision: D46560163

fbshipit-source-id: 1e9fd0858ca3a6392e98c50df89bc7dec8af1735