buck2: move build signals instantiation into with_dice_ctx
Summary:
We want to time the lifetime of build signals to one DICE transaction. There
are a few reasons for this:

- The first one is that now we create them way too early so there is no
  possibility of having any config there.
- The second one is that we would like to be able to access DICE from within
  this listener to e.g. resolve keys (so we can minimize the amount of data the
  listener needs to store), so it must live for less time than DICE (and not
  the opposite).

To do this, we allow creating the "installer" (which we use to inject in user
data) without starting the background thread, which we do once we have access
to DICE.

To make this work, we need to move of the code for BuildSignals outside of
build_api so that it's available in buck2_server_ctx, hence the new
buck2_build_signals crate.

The few `Send` peppered here and there were necessary to un-confuse the
compiler. They should not be actually needed, but they don't seem to harmful.

Reviewed By: bobyangyf

Differential Revision: D46598430

fbshipit-source-id: 882125dd9744f4517e009448d8515c0e56e47c84