Anon targets should support attrs.arg
Summary:
Add `AnonTargetAttr::Arg` use case. Note that query macros are not supported here.

After discussing with cjhopman, it doesn't seem like it's a common use case for anon targets, so we can add it later if someone requests it. There is a bit more complex refactoring to support query macros.

Reviewed By: cjhopman

Differential Revision: D45987654

fbshipit-source-id: bc32c84095bd4f8159aa1bb319cda1e64c1ce9c1