Encode ResolvedStringWithMacros with ConfiguredStringWithMacros
Summary:
We want to support `attrs.arg()` for anon targets. When the `AnonTargetKey` is created, `attrs.arg()`'s value is a `ResolvedStringWithMacros`. We want to be able to use this as part of the DICE key, so `ResolvedStringWithMacros` won't work here. Let's encode it with an optional `ConfiguredStringWithMacros` field, which _does_ work as a DICE key (equality, hashable, static lifetime), and we can resolve it later when we need the resolved `Value<'v>`. `ConfiguredStringWithMacros` is basically a wrapper around `ArcStr`, so these strings are in memory already.

We can add a flag to `attrs.arg()` which specifies if the arg should be compatible with anon targets. If it should be compatible, the `ResolvedStringWithMacros` will be encoded with `ConfiguredStringWithMacros`.

There is a roughly no statistically significant memory increase with this change:
```
absh -irm \
    -a '~/buck2-a --no-buckd --isolation-dir=abc audit providers fb4a --quiet' \
    -b '~/buck2-b --no-buckd --isolation-dir=def audit providers fb4a --quiet'

Time (in seconds):
A: n=24 mean=40.316 std=1.436 se=0.299 min=37.890 max=43.072 med=40.013
B: n=24 mean=40.258 std=1.404 se=0.292 min=38.097 max=43.611 med=40.086
A: distr=[▖    ▗▖▗    ▖  ▖▗ ▖▖▐▖ ▌   ▖  ▖ ▖ ▗   ▗▗▖         ▗▗  ▖     ]
B: distr=[  ▖  ▖▖ ▖ ▗   ▗▖▗ ▖▐▖    ▗▙  ▖      ▗ ▖ ▗ ▐▗               ▗]
B/A: 0.999 0.977..1.020 (95% conf)

Max RSS (in megabytes):
A: n=24 mean=12169 std=78 se=16 min=12022 max=12284 med=12192
B: n=24 mean=12165 std=86 se=18 min=12028 max=12367 med=12165
A: distr=[▖▗ ▖    ▖▖  ▗▗▖ ▗ ▗  ▗  ▐▌▖ ▖▗ ▖▌  ▗▗▗            ]
B: distr=[ ▖ ▗▖▗ ▗  ▖ ▗ ▖   ▄▗▄▗▖▄ ▖ ▗ ▗     ▖  ▗   ▖      ▗]
B/A: 1.000 0.996..1.004 (95% conf)
```

Reviewed By: cjhopman

Differential Revision: D45987650

fbshipit-source-id: e8d4e87a851cba254a21691516619e2eb1bc1dd9