Rework action node caches and don't use temporary_spawn
Summary:
removing the temporary spawns changes the future's lifetime to be no longer static. Therefore, we can no longer just stick the computing future into the cache as a SharedFuture.

Instead, we just do this mechanism where the first to insert into the cache will drive a future to completion, and notify a channel that will hold the result in the cache.
Cancellations will result in restarting the polls with a different future associated with the same node key. However, note that the future being reran is actually a dice task, so really the other callers would be keeping the actual spawned dice work alive, and no true work would actually be reran anyways.

Reviewed By: krallin

Differential Revision: D46503321

fbshipit-source-id: b53a2bf142dba94b3b24dd225b0cf2d1acc1c8e7