targets show output don't use temporary spawn
Summary:
this isn't quite the same as before since the temporary spawns were eager, but the new futures are lazy which means we delay spawning.

but, we always just immediately polled these in a futuresunordered, which would have polled on these futures anyways, so this end up being roughly equivalent.

Reviewed By: stepancheg

Differential Revision: D46503314

fbshipit-source-id: fd7e539192d1994ccaa08d24a7e2c78bdc004ace