Behave like part of the buck2 workspace
Summary:
This `Cargo.toml` is no longer managed by autocargo, so remove the
generated attribution and the `autocargo` field.

`atty` is not part of the workspace, but it is now in the stdlib, so use
that.

Downgrade to `clap-3` because that's what the rest of buck2 uses.

Remove `compatible_with` because with all the platforms listed, it is
redundant.

Reviewed By: krallin

Differential Revision: D46673486

fbshipit-source-id: fa4d1258e3922f3cbddde86eb49fcc58988bae6f