Display for TypeCompiled
Summary:
And use it in error messages.

Note there's some complicated machinery in that `Display for TypeCompiledImplAsStarlarkValue` is not the same as `Display for TypeCompiledImpl`.

It does not matter much now, but if we expose `TypeCompiled` to starlark API, `Display` for starlark type as starlark value `[None]` should be something like `type_compiled([None])` not `[None]`.

Reviewed By: wendy728

Differential Revision: D46624905

fbshipit-source-id: 8f3694bd1741d2e4ad5bce7fea2a34feb06d5fcb