Return structured CxxLinkResult from link methods
Summary:
Follow the pattern of returning a structured result (`CxxLinkResult`) from the link functions. At present, it returns a triple which makes it harder to pass information through the call stack as it requires updating all call sites and it's also hard to document the semantics of the return value at the definition site.

Need driven by adding support for exposing the link command as a subtarget.

Reviewed By: rmaz

Differential Revision: D46681077

fbshipit-source-id: 00771a0fe4898ee620ceb9bd00835bd76e2f1cb9