buck2: java_binary: set allow_cache_upload = True
Summary:
In the top 10 of all actions we spend the most time waiting on, there are
2 java binaries:

- fbcode//fbplatform/gatekeeper/codegen
- fbcode//tools/antlr4

Both of those change approximately never, but they get cache misses all the
time because they are local only.

https://fburl.com/scuba/buck2_critical_path_actions/m0gup48m

Even if we exclude CI those are still fairly significant!

https://fburl.com/scuba/buck2_critical_path_actions/6wjqyije

The critical path potential here is 2.7 days over the course of 3 days, and
that's with only 1/3 of the samples using the backend that emits critical path
improvements at all, so that would mean we spent about 3 days of waiting on
this every day across all our engineers:

https://fburl.com/scuba/buck2_critical_path_actions/0d56y0oz

There is a point being made here in comments about output size causing issues
for RE, but we do have limits on upload size for local uploads, so that seems
fine.

Reviewed By: ndmitchell

Differential Revision: D46720374

fbshipit-source-id: cb7e4c2fcd41722e7493be137e6c82bf54df50ea