Pass a bunch more Android-specific params through the install command
Summary:
Added some more Android-specific params for compatibility with `buck1`. These are again just passed straight through to the installer.

I didn't add all of them, because some of them are incredibly rarely used (if ever) and they do fill up e.g. the help text so it's nicer not to bother with them.

Reviewed By: krallin

Differential Revision: D46686158

fbshipit-source-id: 876047fa19ce9f70706efcb1f06cc1a31e5a1828