Handle preload deps
Summary:
## This stack
See summary of D46476480.

## This diff
Generate shared objects from preload dependencies and all of its transitive dependencies.

These are needed for the iserv script, which is created in the next diff.

## RFC
I'll double check this, but IIUC when I put print statements to debug all this, I noticed that only the dependencies of the preload_deps had their SOs available in the `SharedLibraryInfo` provider.
This is why I ended getting the object files from the preload_deps and manually running the haskell linker to generate SOs from them...

This doesn't sound like' it's the best approach. I would think that at least one of the many providers in each preload_dep should provide the SO that we could just symlink, but I didn't see that when I printed the contents of `preload_dep.providers` in the loop (i.e. printed all the providers of the direct preload_deps).

So the main question is: is this a really bad approach? What am I missing? Is there a better way to get the SOs from preload_deps (e.g. `libcommon_init_kill.so`, `libcommon_memory_jemalloc.so`)?

Reviewed By: simonmar

Differential Revision: D46476949

fbshipit-source-id: 9310c147e809bce218a7716c179c951d9768f375