Surface vpnless support in `buck2 status`
Summary:
As krallin pointed out, adding an e2e test is _way_ easier if you just have buck tell you whether it supports vpnless or not.  Let's do that.

This adds an optional bool field to the status reponse message indicating whether the daemon supports vpnless operation or not. Absense of this boolean means the daemon didn't start (and then the status of vpnless support doesn't really matter)

Reviewed By: krallin

Differential Revision: D46713431

fbshipit-source-id: f56849ad804b7d62ead4e2a8919e7e2abc06c475