buck2: test: put an enum in ExecuteResponse
Summary:
Like it says in the title. I want to be able to return more structured errors
so let's prepare for that.

Note that changing a value to a oneof is not a breaking change in GRPC, though
in reality here it doesn't matter all that much since we release client and
server together.

Reviewed By: get9

Differential Revision: D46799955

fbshipit-source-id: 4581fa0dac1639c7fd4f0b6ccf48090510b00cd5