chore: rework the README a bit (#234)
Summary:
chore: rework the README a bit

 The README as it stands doesn't really "sell" Buck2 very well but more
than that, it barely even expresses what it is or why it exists! It has links,
but generally you should expect people to not follow them if the frontmatter
isn't selling it.

This reworks the frontmatter of the README to drill in three major points about
why Buck2 is interesting as a build system.

- It's Fast(TM)
- It's Hermetic
- It's Multi-Language

There are *other* things that are important too, like what makes it different
from Bazel -- but instead just direct people to look at the "Why Buck?" page on
the site, which goes into more detail. And you don't have unlimited space to
extoll the virtues of the project. So these three seem like the major taglines
to stick too. Plus, Bazel users will likely be willing to look in more detail
since they don't need to be convinced about points 2 and 3, at least.

Furthermore, just point people to the binary releases to get started. For the
vast majority of potential users, there's very little point to listing the
Cargo build instructions when it's not relevant to their immediate interests.

This also moves some of the coding conventions stuff, and the build
instructions, to a new file: `HACKING.md` which is more or less self
explanatory.

Pull Request resolved: https://github.com/facebook/buck2/pull/234

Reviewed By: krallin

Differential Revision: D46833950

Pulled By: ndmitchell

fbshipit-source-id: a307e5ed2d51bbe7a67bd354b27ae7c92d12e622