Always print error messages to stderr
Summary: We sometimes don't manage to print error messages to stderr, which isn't great. We could try and only print to stderr when keep-going is false, but that seems like unnecessary hassle, and also might not be what the user wants. Just switch to always printing to stderr if there is an error.

Reviewed By: krallin

Differential Revision: D46859304

fbshipit-source-id: 8a093d0dc0226cd4bd6b6e728bd9fc7f73c5bae5