Introduce an ExplicitCancellationContext for explicitly cancellable futures
Summary:
Introduce a new type of CancellationContext for ExplicitCancellationFutures that can be used to manage the explicit cancellation of a future different than the existing drop cancel futures.

Currently this doesn't do anything special, but new apis added in D46744416, and used in D46745349.
It also has API to convert itself back to the old CancellationContext for backwards compatibility (added in D46744412)

Update existing critical_section methods, and the like to delegate implementation to the new ExplicitCancellationContext

Reviewed By: krallin

Differential Revision: D46744417

fbshipit-source-id: 8d393aaaeb8a704da719f7ecf6df45a3819918d0