Let cth_tpx state be shared between instances
Summary:
# Context
At the moment cth_tpx will not see failures coming from some callbacks of ct_hooks. This is because even with top priority, it is not always the case that cth_tpx will be the last hook to run in post_... actions: in half of them it will be the first one to run. So in the next diff, we will introduce two version, one with max and one with min priority, so that between the two, they cover all the cases.

# Problem
We will have two versions of the hook running, but we want them to share the underlying state. To do this, we introduce a gen_server that will act as a shared variable, and a single copy of the state will live there.

# This diff
Here we introduce the gen_server and make cth_tpx use it. To minimize code changes, we introduce a combinator `on_shared_state/2` that we can just plug in every CT callback, with the actual state variable, that previously would be the last parameter of the callback now being the single parameter of a lambda.

Reviewed By: TheGeorge

Differential Revision: D46721430

fbshipit-source-id: b59e858a4a47e6e65f41255e7ec75ed9bf38f025