Fix failures in init_per and end_per functions not surfacing as errors
Summary:
When running tests interatively, failures in init_per_* and end_per_* functions would not be surfaced as failures. However, failures in post_init_per_* and post_end_per_* hooks were being surfaced correctly

The correct protocol for hooks to turn override a success into a failure is by returning a `{fail, xxxx}` term. We had the proper logic in place to thread this terms around and interpret them as failures. However, the "base case" seemed to be wrong: when an init_per or end_per function failed, we'd generate a `{error, xxxxx}` term instead.

We now return a `{fail, xxxx}` term instead, by changing `ct_daemon_hooks:format_ct_error/3`. Because this function seems to be the only place where a `thrown` atom is introduced, we also change `ct_daemon_printer:printe_result/2`, that seems to be the only consumer, but I'm unsure if this makes any difference.

Reviewed By: TheGeorge

Differential Revision: D46894301

fbshipit-source-id: e39e4474d9bb6b9bc27dac5fc6de9cdc11475188