buck2: buck2_wrapper_common: expose a way to get the status of a process
Summary:
I want to use this for debugging. This isn't a great implementation. It uses
the sysinfo crate, which has a very byzantine API, but it *is* portable,
and I'd rather not reimplement this for Mac in particular where I haven't been
able to find any docs on the underlying `proc_pidinfo` API.

Reviewed By: wendy728

Differential Revision: D46934017

fbshipit-source-id: aa2bed31fa358f3947845e24d58d141ed102f46d