buck2: kill: if you time out connecting to the daemon, kill it
Summary:
Right now, if you SIGSTOP your Buck2 daemon, then when you try to kill it,
we happily claim there's no daemon.

That's completely wrong, and if you try to run a command later you are stuck.
This is bad!

To fix this, let's have `kill` actually kill the daemon in this case. We don't
want to kill just any PID we find out our PID file, so we guard against this by
only killing if we get a timeout connecting, which means we probably didn't get
a connection refused.

Reviewed By: wendy728

Differential Revision: D46934016

fbshipit-source-id: c3d490d7d350eb82ce26d678120182feb0b7deb4