buck2: use lower level methods in `kill`
Summary:
My end goal here is to have `buck2 kill` handle the case where we can read
the process info, and then we time out connecting, and have that result in
us killing the daemon.

To do that, I need to separate those steps (because I don't want to potentially
kill a daemon that is not the one I tried to connect to).

This is what this does. This also moves `kill` to not require `with_flushing()`
because that wasn't actually doing anything anyway.

Reviewed By: wendy728

Differential Revision: D46934021

fbshipit-source-id: e0d6d71dcec19af108dd53d1268c3fbcb503dacf