buck2: Eden I/O: migrate to getAttributesFromFilesV2
Summary:
Unlike getAttributesFromFiles it is not deprecated. The reason I'm doing this
now is because I would like to add support for reading Blake3 hashes out of
Eden, and this is only implemented in getAttributesFromFilesV2.

This was introduced in D38090796, though I don't know the exact version it
rolled out in. I picked a version from September later that year to give
ourselves a bit of headroom.

Reviewed By: themarwhal

Differential Revision: D46798142

fbshipit-source-id: f37f5c8d18ab0e3ecaeee98d8ed61498190336eb