Remove return_type from cmd_args builder
Summary:
I want to remove all string type annotation. At least, try.

The idea is this.

Currently type annotations are passed to `Ty::parse`:

https://www.internalfb.com/code/fbsource/[ede43e72b574a3304a85d83352016d5906712c9e]/fbcode/buck2/starlark-rust/starlark_derive/src/module/render/mod.rs?lines=235

There are two issues here. First, we cannot access globals which are needed so we could reference types like `list` (not `"list"`).

Second, `Ty::parse` calls `Ty::from_expr`, so `Ty::from_expr` cannot use `CstPayload`. So when `Ty::from_expr` is called from typechecker, it cannot access resolved globals which are stored in `CstPayload` (which are supposed to be resolved even in lint typechecker), so again we could properly resolve `list` (not `"list"`).

Various workarounds possible, but the straigthforward fix seems to be just not use string type expressions in native functions.

With this diff, we have only 9 uses of `return_type` in `app/` (and some number of type expressions for parameters).

Reviewed By: ndmitchell

Differential Revision: D46911981

fbshipit-source-id: a932680150ccc8e141d1f71d2ab73d5697db772b