worker.proto: Add cancellable grpc execute method
Summary:
Using streaming request, similar to forkserver.

This is a breaking change so the plan is to make javacd/kotlincd support both the existing `Execute` and the new cancellable method, update buck2 and then remove the unused method.

Calling it `Exec` since `ExecuteCommand` is taken and `ExecuteCancellable` is a bit wordy/redundant if it's the only `Execute` method.

Reviewed By: IanChilds

Differential Revision: D47010918

fbshipit-source-id: c041834d851eb33e3cebb47aca58ca5844f4336b