Add new api on explicit cancellable future for managing prevent cancellation
Summary:
New API that offers a more flexible way of managing the disabling of cancellations during a future.
This now lets you acquire a token that as long as it's held, prevents cancellations.
This means one can acquire this head of the actual future execution, and drop/release it at any point to allow cancellations again. Furthermore, this can be converted to disabling cancellations forever easily.

Used by modern dice in D46745349

Reviewed By: krallin

Differential Revision: D46744416

fbshipit-source-id: 068f9f277ab230aeb735e7eb5a0fc2461e32ab65