Displayed network stats are now deltas
Summary:
Currently, network stats (RE and HTTP client) are being displayed as counters over the entire daemon lifetime. This is misrespresentative and they should be reported with respect to the command start, not the daemon.

Superconsole now displays counters that are a delta: `network stats during first Snapshot` - `network stats during last Snapshot`.

I overlooked doing this earlier as a follow-up to D46733525, sorry.

Reviewed By: krallin

Differential Revision: D47082380

fbshipit-source-id: 24ea3ff2500ae058d482b85316a6d7b065d14cfa