Add audit args truncation
Summary:
We are dropping invocation recorder 600-700 times per hour due to oversize https://fburl.com/logview/6k09taf3. Which means the buck2_builds record is not available for those builds. Current major source of this is Audit command which comes with a long args field executed by SC.

This change is adding a truncation for it.

Reviewed By: JakobDegen

Differential Revision: D47101274

fbshipit-source-id: 0d3f38429809a49fc429f38e36a43eabf7c6c2d2