chore(nix): update flake, add a few packages to `nix develop` shell (#314)
Summary:
This adds `reindeer`, `mold`, `clang` and `lld` to the shell environment provided by `nix develop`. These packages are enough to perform a self bootstrap of buck2-with-buck2 out of the box. `clang` and `lld` are needed as the default cxx toolchain and rust builds will mandate them.

The flake update is required to get a newer copy of `reindeer` from nixpkgs, which can correctly generate shim files capable of handling krallin's fork of `perf-event` while `vendor = false`.

The inclusion of `mold` is not because I'm using it with buck2-on-buck2, but because in debug builds with `cargo`, re-linking buck2 is slower than I'd like, and that can be fixed with a tweak to the `.cargo/config.toml` file (which I might file separately.)

Pull Request resolved: https://github.com/facebook/buck2/pull/314

Reviewed By: stepancheg

Differential Revision: D47146463

Pulled By: ndmitchell

fbshipit-source-id: 70217b5519cb16565066259dfdfe50b953dbd08f